requirements.txt
appdirs==1.4.3
asgiref==3.3.1
astroid==2.5.1
autopep8==1.5.5
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.7
django-admin-interface==0.15.5
django-colorfield==0.4.1
django-environ==0.4.5
django-flat-responsive==2.0
django-flat-theme==1.1.4
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
isort==5.7.0
lazy-object-proxy==1.5.2
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
psycopg2==2.8.6
pycodestyle==2.7.0
pylint==2.7.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
pyparsing==2.4.6
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.8
waitress==2.0.0
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
